+++ packages/app/package.json
@@ -14,41 +14,21 @@
     "lint": "backstage-cli package lint"
   },
   "dependencies": {
-    "@backstage-community/plugin-github-actions": "^0.6.16",
-    "@backstage-community/plugin-tech-radar": "^0.7.4",
-    "@backstage/app-defaults": "^1.7.9",
-    "@backstage/catalog-model": "^1.9.0",
     "@backstage/cli": "^0.36.3",
-    "@backstage/core-app-api": "^1.20.2",
     "@backstage/core-components": "^0.18.11",
-    "@backstage/core-plugin-api": "^1.12.7",
+    "@backstage/frontend-defaults": "^0.3.3",
+    "@backstage/frontend-plugin-api": "^0.17.2",
     "@backstage/integration-react": "^1.2.19",
-    "@backstage/plugin-api-docs": "^0.14.2",
-    "@backstage/plugin-catalog": "^2.0.6",
-    "@backstage/plugin-catalog-common": "^1.1.10",
-    "@backstage/plugin-catalog-graph": "^0.6.5",
-    "@backstage/plugin-catalog-import": "^0.13.14",
-    "@backstage/plugin-catalog-react": "^3.1.0",
-    "@backstage/plugin-org": "^0.7.5",
-    "@backstage/plugin-permission-react": "^0.5.2",
-    "@backstage/plugin-scaffolder": "^1.38.0",
+    "@backstage/plugin-app-react": "^0.2.4",
     "@backstage/plugin-search": "^1.7.5",
-    "@backstage/plugin-search-react": "^1.11.5",
-    "@backstage/plugin-techdocs": "^1.17.7",
-    "@backstage/plugin-techdocs-module-addons-contrib": "^1.1.37",
-    "@backstage/plugin-techdocs-react": "^1.3.12",
     "@backstage/plugin-user-settings": "^0.9.4",
-    "@backstage/theme": "^0.7.3",
     "@backstage/ui": "^0.16.0",
     "@material-ui/core": "^4.12.2",
     "@material-ui/icons": "^4.9.1",
     "@parsifal-m/backstage-plugin-open-feedback": "workspace:^",
-    "history": "^5.0.0",
     "react": "^18.0.2",
     "react-dom": "^18.0.2",
-    "react-router": "^6.3.0",
-    "react-router-dom": "^6.3.0",
-    "react-use": "^17.2.4"
+    "react-router-dom": "^6.3.0"
   },
   "devDependencies": {
     "@backstage/test-utils": "^1.7.19",
