+++ packages/app/src/index.tsx
@@ -1,6 +1,6 @@
 import '@backstage/cli/asset-types';
 import '@backstage/ui/css/styles.css';
 import ReactDOM from 'react-dom/client';
-import App from './App';
+import app from './App';
 
-ReactDOM.createRoot(document.getElementById('root')!).render(<App />);
+ReactDOM.createRoot(document.getElementById('root')!).render(app);
